Configure upload/download bandwidth separately
Co-authored with @beenanner.

The Traffic Control tool (tc) can only classify and police an
interface's ingress traffic, e.g. dropping packets that exceed some
rate. However, one of those policing actions is to redirect classified
traffic to another device. By shaping the egress traffic on the latter
device, we shape ingress traffic on the former. Use the Intermediate
Functional Block (IFB) loadable kernel module to create such a device,
a virtual one.

Note that, currently, we simply shape the traffic on the IFB device as
we shape the traffic on the physical device.

See sitespeedio/sltc#1, phuedx/sltc#1, and phuedx/sltc#3 for additional
detail and discussion.

Closes sitespeedio/sltc#1.